+++ src/components/BulkEditList/BulkEditListResult/NoResultsMessage/NoResultsMessage.js
@@ -3,7 +3,7 @@ import { useLocation } from 'react-router';
 import { FormattedMessage } from 'react-intl';
 import { Layout, Icon } from '@folio/stripes/components';
 
-import { translationSuffix } from '../../../../constants';
+import { CRITERIA, translationSuffix } from '../../../../constants';
 
 import css from './NoResultsMessage.css';
 
@@ -11,12 +11,15 @@ export const NoResultsMessage = () => {
   const location = useLocation();
   const search = new URLSearchParams(location.search);
   const capabilities = search.get('capabilities');
+  const criteria = search.get('criteria');
 
   const message = useMemo(() => {
     const identifier = new URLSearchParams(location.search).get('identifier');
     const messagePrefix = identifier ? `.${identifier}` : '';
 
-    return <FormattedMessage id={`ui-bulk-edit.list.result.emptyMessage${translationSuffix[capabilities]}${messagePrefix}`} />;
+    if (criteria === CRITERIA.LOGS) {
+      return <FormattedMessage id="ui-bulk-edit.list.result.emptyMessage.logs" />;
+    } else return <FormattedMessage id={`ui-bulk-edit.list.result.emptyMessage${translationSuffix[capabilities]}${messagePrefix}`} />;
   }, [location.search]);
 
   return (